+++ user/bus_drv/bus_rs485.h
@@ -40,6 +40,8 @@ bool CheckA308M();
 bool CheckUltraSound();
 bool CheckA6D6();
 bool CheckCDMCo2();
+bool CheckCDMPm25();
+bool CheckCDMTvoc();
 bool CheckBtmG6();
 bool GetCustomSerial();